Form>>distinguish: aName
We can enforce the only one distinction per name restriction because the first axiom of Laws of Form says that invoking the same name twice is the same as invoking it once. In other words, we are not altering the meaning of the form structure by doing so. To that effect, pay close attention at how ifNil: is not being used as a control mechanism, but rather as a means of replacement. Indeed, ifNil: should be read as meaning butIfNil: instead.